+++ src/main/java/com/skycat/mystical/ConfigModel.java
@@ -42,6 +42,7 @@ public class ConfigModel {
     @Nest public SoundSwapConfig soundSwap = new SoundSwapConfig();
     @Nest public FishingRodSwapConfig fishingRodSwap = new FishingRodSwapConfig();
     @Nest public MysteryEggsConfig mysteryEggs = new MysteryEggsConfig();
+    @Nest public EvilBeesConfig evilBees = new EvilBeesConfig();
 
     @SectionHeader("Logging") // Note: Logging as ERROR level does not always mean a critical error.
     public LogLevel failedToSetNightTimerLogLevel = LogLevel.WARN;
@@ -423,6 +424,16 @@ public class ConfigModel {
         public static boolean chancePredicate(double value) {
             return ConfigModel.chancePredicate(value);
         }
+        public static boolean weightPredicate(double value) {
+            return ConfigModel.weightPredicate(value);
+        }
+    }
+
+    public static class EvilBeesConfig {
+        public boolean enabled = true;
+        public LogLevel logLevel = LogLevel.OFF;
+        @PredicateConstraint("weightPredicate")
+        public double weight = 1;
         public static boolean weightPredicate(double value) {
             return ConfigModel.weightPredicate(value);
         }